+++ Proyecto/src/proyectofinal/vista/Cliente.java
@@ -1,0 +1,9 @@
+package proyectofinal.vista;
+
+/**
+ *
+ * @author SPARTAN PC
+ */
+public class Cliente extends Usuario {
+    
+}
