+++ ProyectoFinal/Prueba2.csv
@@ -1,0 +1,10 @@
+0,Comida nombre,0.0,Comida descripcion,10
+1,Comida nombre,100.0,Comida descripcion,11
+2,Comida nombre,200.0,Comida descripcion,12
+3,Comida nombre,300.0,Comida descripcion,13
+4,Comida nombre,400.0,Comida descripcion,14
+5,Comida nombre,500.0,Comida descripcion,15
+6,Comida nombre,600.0,Comida descripcion,16
+7,Comida nombre,700.0,Comida descripcion,17
+8,Comida nombre,800.0,Comida descripcion,18
+9,Comida nombre,900.0,Comida descripcion,19
