+++ ProyectoFinal/Prueba3.csv
@@ -1,0 +1,10 @@
+1,Chapulines,200.0,100g Preparados con Ajo y Chile de Árbol servidos con Guacamole con quesillo y tortillas,60
+2,Tostadas de Atún Ahumado (3pzas),290.0,A la Mexicana Servidos con Lechuga Cremoso de Aguacate Sobre Tostada de Maíz Azul,52
+3,Pambazos de Short Rib (3pzas),265.0,Rellenos de Papa y Bañados en Salsa de Guajillo Acompañados de Lechuga Queso de Ocosingo y Crema,95
+4,Clásico Caldo Tlalpeño,190.0,De Pollo con Caldo de Jitomate Pasilla Mixe y Chipotle Arroz Blanco Vegetales Aguacate y Quesillo,86
+5,Huachinango con Mole Chichilo,260.0,200g con Mole Chichilo Puré de Plátano Chayote Cambray y Ejotes,72
+6,Robalo al pastor con Costra de Huitlacoche,250.0,Piña Tatemada Puré de Piña y Mayonesa de Ajo,51
+7,Tacos Rib Eye (3pzas),250.0,250g con Costra de Queso y Rajas de Chile Poblano,62
+8,Filete de Res con Costra de Pepitas de Calabaza,280.0,Calabaza Támala Tatemada Puré de Calabaza y Salsa de Flor de Calabaza85,0
+9,Buñuelo con Mousse de Quesillo,170.0,Con Almendras Garapiñadas y en Polvo Merengue Deshidratado y Granizado de Café de Olla,42
+10,Delicia de Xocolatl (1 rebanada),200.0,Pastel de Chocolate de Metate con Helado de Fresa y Betabel con Mousse de Maracuyá y Aceite de Poleo en Polvo,21
