+++ ProyectoFinal/nbproject/private/private.properties
@@ -1,0 +1,2 @@
+compile.on.save=true
+user.properties.file=C:\\Users\\alanm\\AppData\\Roaming\\NetBeans\\12.6\\build.properties
