--- src/proyectofinal/vista/Cliente.java
@@ -1,9 +1,0 @@
-package proyectofinal.vista;
-
-/**
- *
- * @author SPARTAN PC
- */
-public class Cliente extends Usuario {
-    
-}
